refactor other helpers into helpers lib, consolidate
